test: keep State getter coverage in unit suite
